# packages/api-client/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Public REST API of Honest Robin: Time. Authenticate with a personal access token (`Authorization: Bearer hrt_…`). Session-authenticated browser clients send the `HonestRobin-Account-Id` header to pick an account.",
    "license" : {
      "name" : "AGPL-3.0-only",
      "url" : "https://www.gnu.org/licenses/agpl-3.0.html"
    },
    "title" : "Honest Robin: Time API",
    "version" : "v1"
  },
  "security" : [ {
    "token" : [ ]
  } ],
...
  "paths" : {
...
    "/api/v1/expenses" : {
      "get" : {
        "operationId" : "expense_list",
        "parameters" : [ {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "in" : "query",
          "name" : "membership_id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "query",
          "name" : "project_id",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "query",
          "name" : "cursor",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageExpenseView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "expenses" ]
      },
      "post" : {
        "operationId" : "expense_create",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ExpenseInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ExpenseView"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "expenses" ]
      }
    },
...
    "/api/v1/reports/budget" : {
      "get" : {
        "operationId" : "report_budget",
        "parameters" : [ {
          "in" : "query",
          "name" : "client_id",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }, {
          "in" : "query",
          "name" : "include_archived",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BudgetReport"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Budget, spent, remaining and % used per project",
        "tags" : [ "reports" ]
      }
    },
    "/api/v1/reports/detailed" : {
      "get" : {
        "operationId" : "report_detailed",
        "parameters" : [ {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "in" : "query",
          "name" : "client_id",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }, {
          "in" : "query",
          "name" : "project_id",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }, {
          "in" : "query",
          "name" : "task_id",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }, {
          "in" : "query",
          "name" : "membership_id",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }, {
          "in" : "query",
          "name" : "team_id",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }, {
          "in" : "query",
          "name" : "billable",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "in" : "query",
          "name" : "approval_state",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }, {
          "in" : "query",
          "name" : "invoiced",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "in" : "query",
          "name" : "cursor",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageTimeEntryView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Time entries, oldest first, with the time report's filters",
        "tags" : [ "reports" ]
      }
    },
    "/api/v1/reports/expenses" : {
      "get" : {
        "operationId" : "report_expenses",
        "parameters" : [ {
          "in" : "query",
          "name" : "group_by",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "category"
          }
        }, {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "in" : "query",
          "name" : "client_id",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }, {
          "in" : "query",
          "name" : "project_id",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }, {
          "in" : "query",
          "name" : "membership_id",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }, {
          "in" : "query",
          "name" : "team_id",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }, {
          "in" : "query",
          "name" : "category_id",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }, {
          "in" : "query",
          "name" : "billable",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "in" : "query",
          "name" : "approval_state",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }, {
          "in" : "query",
          "name" : "invoiced",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ExpenseReport"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Expenses grouped by category, client, project or person",
        "tags" : [ "reports" ]
      }
    },
    "/api/v1/reports/time" : {
      "get" : {
        "operationId" : "report_time",
        "parameters" : [ {
          "in" : "query",
          "name" : "group_by",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "project"
          }
        }, {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "in" : "query",
          "name" : "client_id",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }, {
          "in" : "query",
          "name" : "project_id",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }, {
          "in" : "query",
          "name" : "task_id",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }, {
          "in" : "query",
          "name" : "membership_id",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }, {
          "in" : "query",
          "name" : "team_id",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }, {
          "in" : "query",
          "name" : "billable",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "in" : "query",
          "name" : "approval_state",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }, {
          "in" : "query",
          "name" : "invoiced",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TimeReport"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Hours and amounts grouped by client, project, task or person",
        "tags" : [ "reports" ]
      }
    },
    "/api/v1/reports/uninvoiced" : {
      "get" : {
        "operationId" : "report_uninvoiced",
        "parameters" : [ {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "in" : "query",
          "name" : "client_id",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }, {
          "in" : "query",
          "name" : "project_id",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UninvoicedReport"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Billable time and expenses not invoiced yet, by client and project",
        "tags" : [ "reports" ]
      }
    },
    "/api/v1/reports/{kind}/export" : {
      "get" : {
        "operationId" : "report_export",
        "parameters" : [ {
          "in" : "path",
          "name" : "kind",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "format",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "csv"
          }
        }, {
          "in" : "query",
          "name" : "group_by",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "in" : "query",
          "name" : "client_id",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }, {
          "in" : "query",
          "name" : "project_id",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }, {
          "in" : "query",
          "name" : "task_id",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }, {
          "in" : "query",
          "name" : "membership_id",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }, {
          "in" : "query",
          "name" : "team_id",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }, {
          "in" : "query",
          "name" : "category_id",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }, {
          "in" : "query",
          "name" : "billable",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "in" : "query",
          "name" : "approval_state",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }, {
          "in" : "query",
          "name" : "invoiced",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "in" : "query",
          "name" : "include_archived",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "summary" : "Export a report as CSV or XLSX",
        "tags" : [ "reports" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "BudgetReport" : {
        "properties" : {
          "rows" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/BudgetReportRow"
            }
          }
        },
        "required" : [ "rows" ]
      },
      "BudgetReportRow" : {
        "properties" : {
          "budget" : {
            "$ref" : "#/components/schemas/BudgetView"
          },
          "client" : {
            "$ref" : "#/components/schemas/Ref"
          },
          "project" : {
            "$ref" : "#/components/schemas/ProjectRef"
          }
        },
        "required" : [ "budget", "client", "project" ]
      },
...
      "CurrencyAmount" : {
        "properties" : {
          "billable_amount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "currency" : {
            "type" : "string"
          },
          "uninvoiced_amount" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "billable_amount", "currency", "uninvoiced_amount" ]
      },
...
      "ExpenseAmount" : {
        "properties" : {
          "amount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "billable_amount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "currency" : {
            "type" : "string"
          },
          "uninvoiced_amount" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "amount", "billable_amount", "currency", "uninvoiced_amount" ]
      },
...
      "ExpenseInput" : {
        "properties" : {
          "amount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "billable" : {
            "type" : "boolean"
          },
          "category_id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "membership_id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "notes" : {
            "type" : "string"
          },
          "project_id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "spent_date" : {
            "type" : "string",
            "format" : "date"
          },
          "units" : {
            "type" : "number"
          }
        }
      },
      "ExpenseReport" : {
        "properties" : {
          "amounts" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ExpenseAmount"
            }
          },
          "count" : {
            "type" : "integer",
            "format" : "int32"
          },
          "from" : {
            "type" : "string",
            "format" : "date"
          },
          "group_by" : {
            "type" : "string"
          },
          "rows" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ExpenseReportRow"
            }
          },
          "to" : {
            "type" : "string",
            "format" : "date"
          }
        },
        "required" : [ "amounts", "count", "group_by", "rows" ]
      },
      "ExpenseReportRow" : {
        "properties" : {
          "amounts" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ExpenseAmount"
            }
          },
          "client_name" : {
            "type" : "string"
          },
          "count" : {
            "type" : "integer",
            "format" : "int32"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          }
        },
        "required" : [ "amounts", "count", "id", "name" ]
      },
      "ExpenseView" : {
        "properties" : {
          "amount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "approval_state" : {
            "type" : "string"
          },
          "billable" : {
            "type" : "boolean"
          },
          "category" : {
            "$ref" : "#/components/schemas/Ref"
          },
          "client" : {
            "$ref" : "#/components/schemas/Ref"
          },
          "created_at" : {
            "type" : "string",
            "format" : "date-time"
          },
          "currency" : {
            "type" : "string"
          },
          "has_receipt" : {
            "type" : "boolean"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "invoice_id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "is_locked" : {
            "type" : "boolean"
          },
          "notes" : {
            "type" : "string"
          },
          "person" : {
            "$ref" : "#/components/schemas/Ref"
          },
          "project" : {
            "$ref" : "#/components/schemas/ProjectRef"
          },
          "receipt_filename" : {
            "type" : "string"
          },
          "spent_date" : {
            "type" : "string",
            "format" : "date"
          },
          "units" : {
            "type" : "number"
          },
          "updated_at" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "amount", "approval_state", "billable", "category", "client", "created_at", "currency", "has_receipt", "id", "is_locked", "person", "project", "spent_date", "updated_at" ]
      },
...
      "PageExpenseView" : {
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ExpenseView"
            }
          },
          "next_cursor" : {
            "type" : "string"
          }
        },
        "required" : [ "data" ]
      },
...
      "PageTimeEntryView" : {
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TimeEntryView"
            }
          },
          "next_cursor" : {
            "type" : "string"
          }
        },
        "required" : [ "data" ]
      },
...
      "ProjectRef" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          }
        },
        "required" : [ "id", "name" ]
      },
...
      "Ref" : {
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          }
        },
        "required" : [ "id", "name" ]
      },
...
      "TimeReport" : {
        "properties" : {
          "amounts" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CurrencyAmount"
            }
          },
          "billable_seconds" : {
            "type" : "integer",
            "format" : "int64"
          },
          "cost_amount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "cost_currency" : {
            "type" : "string"
          },
          "entry_count" : {
            "type" : "integer",
            "format" : "int32"
          },
          "from" : {
            "type" : "string",
            "format" : "date"
          },
          "group_by" : {
            "type" : "string"
          },
          "rounding_minutes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "rounding_mode" : {
            "type" : "string"
          },
          "rows" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TimeReportRow"
            }
          },
          "seconds" : {
            "type" : "integer",
            "format" : "int64"
          },
          "to" : {
            "type" : "string",
            "format" : "date"
          }
        },
        "required" : [ "amounts", "billable_seconds", "cost_amount", "cost_currency", "entry_count", "group_by", "rounding_minutes", "rounding_mode", "rows", "seconds" ]
      },
...
      "UninvoicedReport" : {
        "properties" : {
          "clients" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UninvoicedClientRow"
            }
          },
          "from" : {
            "type" : "string",
            "format" : "date"
          },
          "to" : {
            "type" : "string",
            "format" : "date"
          },
          "totals" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UninvoicedTotals"
            }
          }
        },
        "required" : [ "clients", "totals" ]
      },
...
    },
    "securitySchemes" : {
      "token" : {
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}